requirements.txt
aiohttp>=3.9.0
py-clob-client>=0.18.0
cryptography>=42.0.0
python-dotenv>=1.0.0
web3>=6.0.0
